+++ elgg-plugin.php
@@ -1,6 +1,7 @@
 <?php
 
 return [
+	'bootstrap' => \hypeJunction\Invite\Bootstrap::class,
 	'entities' => [
 		[
 			'type' => 'object',
@@ -35,6 +36,7 @@ return [
 			'resource' => 'friends/invite',
 			'middleware' => [
 				\Elgg\Router\Middleware\Gatekeeper::class,
+				\Elgg\Router\Middleware\UserPageOwnerGatekeeper::class,
 			],
 		],
 		'invite:request' => [
@@ -64,12 +66,43 @@ return [
 			'controller' => \hypeJunction\Invite\InviteGroupMembersAction::class,
 		],
 	],
-
 	'events' => [
-		'seeds' => [
-			'database' => [
-				[\hypeJunction\Invite\Seeder::class, 'addSeed'] => [],
+		'registration_url' => [
+			'site' => [
+				\hypeJunction\Invite\GenerateRegistrationUrl::class => [],
 			],
+		],
+		'register' => [
+			'user' => [
+				\hypeJunction\Invite\RegistrationForwardUrl::class => ['priority' => 1000],
+			],
+			'menu:page' => [
+				\hypeJunction\Invite\PageMenu::class => [],
+			],
+			'menu:entity' => [
+				\hypeJunction\Invite\EntityMenu::class => [],
+			],
+		],
+		'accept' => [
+			'invite' => [
+				\hypeJunction\Invite\AcceptFriendInvitesOnRegistration::class => [],
+				\hypeJunction\Invite\AcceptGroupInvitesOnRegistration::class => [],
+			],
+		],
+		'create' => [
+			'user' => [
+				\hypeJunction\Invite\ProcessUserInvitesOnRegistration::class => [],
+			],
+		],
+	],
+	'view_extensions' => [
+		'register/extend' => [
+			'forms/register/invitation_code' => ['priority' => 100],
+		],
+	],
+	'group_tools' => [
+		'invites' => [
+			'default_on' => false,
 		],
 	],
 ];