+++ views/default/object/user_invite_request/actions.mjs
@@ -1,5 +1,5 @@
+import 'jquery';
 import Ajax from 'elgg/Ajax';
-import $ from 'jquery';
 
 $(document).on('click', '.elgg-menu-user-invite-request > li > a', function (e) {
 	e.preventDefault();